Move optional into module
